backend/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Gorsthir",email = "194412152+G0rsthir@users.noreply.github.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.135.1,<1.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "typer (>=0.24.1,<1.0.0)",
    "uvicorn (>=0.41.0,<1.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "aiosqlite (>=0.22.1,<1.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-multipart (>=0.0.22,<1.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "pypdf (>=6.9.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
server = 'server.cli:cli'


[tool.ruff]
# Allow lines to be as long as 120.
line-length = 120

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # flake8-bugbear
    "B",
    # pyupgrade
    "UP",
    # isort
    "I",
]

ignore = [
    # line too long
    "E501",
    # do not perform function calls in argument defaults
    "B008",
    # type hint typing.Optional
    "UP007"
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pydocstyle]
convention = "google"
